[リファクタリング] MainControllerのJsonやWebSocketの情報を削除した
MainControllerがサーバーへメッセージを送る際、JsonやWebSocketの情報を知らなくて済むようになった
