+++ Client/Assets/RollingBall.unity
@@ -736,6 +736,7 @@ GameObject:
   - component: {fileID: 2134770748}
   - component: {fileID: 2134770749}
   - component: {fileID: 2134770750}
+  - component: {fileID: 2134770751}
   m_Layer: 0
   m_Name: Main
   m_TagString: Untagged
@@ -759,6 +760,7 @@ MonoBehaviour:
   otherPlayerPrefab: {fileID: 1000013530691402, guid: 8da97395e89154d00aeaf2c9abf021e2,
     type: 2}
   itemPrefab: {fileID: 1800889081646452, guid: 94eca55f9d12f4d9eb776b8feae42ae5, type: 2}
+  messageSenderToServer: {fileID: 2134770751}
 --- !u!4 &2134770747
 Transform:
   m_ObjectHideFlags: 0
@@ -805,5 +807,16 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 951bacae69aa3624fbd563d44b902c00, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &2134770751
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 2134770745}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: c49d292764ecea648bbf4bfbaafdd294, type: 3}
   m_Name: 
   m_EditorClassIdentifier:
