+++ Client/Assets/Scripts/MessageSenderToServer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c49d292764ecea648bbf4bfbaafdd294
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
